Settings re-org
