+++ components/screens/settings/Content/ReadSettingsScreen.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { useTheme, VStack } from "native-base";
 import { Switch } from "react-native";
-import CTable from "../../ui/table/CTable";
-import CSection from "../../ui/table/CSection";
-import CCell from "../../ui/table/CCell";
-import { useAppDispatch, useAppSelector } from "../../../store";
-import { selectSettings } from "../../../slices/settings/settingsSlice";
-import { setSetting } from "../../../slices/settings/settingsActions";
+import CTable from "../../../ui/table/CTable";
+import CSection from "../../../ui/table/CSection";
+import CCell from "../../../ui/table/CCell";
+import { useAppDispatch, useAppSelector } from "../../../../store";
+import { selectSettings } from "../../../../slices/settings/settingsSlice";
+import { setSetting } from "../../../../slices/settings/settingsActions";
 
 function ReadSettingsScreen() {
   const theme = useTheme();
